requirements.txt
PyPDF2
python-docx
tk
bcrypt==3.2.1

